+++ test_hevy.py
@@ -1,5 +1,0 @@
-from integrations.hevy import fetch_and_save_workouts
-
-workouts = fetch_and_save_workouts()
-
-print(workouts)